# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: PH Shoes — Alerts Admin/Web API
  version: 1.0.0
servers:
  - url: /api/v1

tags:
  - name: Alerts
    description: Alerts CRUD endpoints for authenticated users.
  - name: AlertsScheduler
    description: Internal endpoints to observe/trigger the alerts scheduler.
paths:
  /alerts-scheduler/run:
    get:
      tags: [AlertsScheduler]
      summary: Run alerts scheduler for a given date
      description: Optional email limits execution to the account owner for dry/test runs.
      operationId: runAlertsScheduler
      security: []
      parameters:
        - in: query
          name: date
          required: false
          schema:
            type: string
            format: date
        - in: query
          name: email
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Scheduler run summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertsSchedulerRunResponse'
  /alerts:
    post:
      tags: [Alerts]
      summary: Create an alert for the authenticated user
      operationId: createAlert
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertResponse'
    get:
      tags: [Alerts]
      summary: List alerts for the authenticated user
      operationId: listAlerts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AlertResponse'

  /alerts/{productId}:
    get:
      tags: [Alerts]
      summary: Get an alert by productId for the authenticated user
      operationId: getAlert
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertResponse'
        '404':
          description: Not Found
    put:
      tags: [Alerts]
      summary: Update an alert (criteria or reset status)
      operationId: updateAlert
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertResponse'
        '404':
          description: Not Found
    delete:
      tags: [Alerts]
      summary: Delete an alert
      operationId: deleteAlert
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Deleted

components:
  schemas:
    AlertCreateRequest:
      $ref: 'components/schemas/Alert.yaml#/components/schemas/AlertCreateRequest'
    AlertUpdateRequest:
      $ref: 'components/schemas/Alert.yaml#/components/schemas/AlertUpdateRequest'
    AlertResponse:
      $ref: 'components/schemas/Alert.yaml#/components/schemas/AlertResponse'
    AlertStatus:
      $ref: 'components/schemas/Alert.yaml#/components/schemas/AlertStatus'
    AlertChannel:
      $ref: 'components/schemas/Alert.yaml#/components/schemas/AlertChannel'
    AlertsSchedulerRunResponse:
      $ref: 'components/schemas/AlertsSchedulerRun.yaml#/components/schemas/AlertsSchedulerRunResponse'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT